pyproject.toml
# pyproject.toml (Updated)

[build-system]
# Specifies build tools. setuptools is common.
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# --- Core Metadata ---
name = "mbkauthepy" # The name used for pip install (MUST be unique on PyPI)
version = "1.6.1" # Your current version
authors = [
  { name="Maaz.waheed", email="maaz.waheed@mbktechstudio.com" },
  # Add others if applicable
]
description = "A fully featured, secure, and extensible authentication system for Python Flask applications.Originally ported from the Node.js version to provide multi-language support for full-stack apps."
readme = "README.md" # Specifies the README file for the long description on PyPI
requires-python = ">=3.8" # Minimum Python version supported
# Using SPDX identifier is preferred over license file for modern tooling, but file is okay too.
# license = { text = "MPL-2.0" } # Example using SPDX identifier
license = { file = "LICENSE" } # Points to your LICENSE file (e.g., MPL-2.0)
keywords = ["flask", "authentication", "session", "auth", "postgresql", "security"]
classifiers = [
    # Trove classifiers - help users find your package
    "Development Status :: 3 - Alpha", # Or 4 - Beta, 5 - Production/Stable
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration :: Authentication/Directory",
    "Topic :: Security",
    "Framework :: Flask",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)", # Keep if license file is MPL-2.0
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Web Environment",
]

# --- Dependencies ---
# List packages your library *requires* to run
dependencies = [
    "Flask>=2.0",
    "Flask-Session>=0.5",
    "psycopg2-binary>=2.9", # Or psycopg2 if users compile
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
    "requests>=2.28",
    "pyotp>=2.8",
    "Flask-Cors>=4.0", # If CORS is integral
    "SQLAlchemy>=1.4", # Needed if using sqlalchemy session type internally
    "importlib-metadata; python_version<'3.10'", # Conditional dependency
     "pybars3",
]

# --- Project URLs ---
# Links displayed on PyPI
[project.urls]
Homepage ="https://github.com/42Wor/mbkauthepy"
Repository ="https://github.com/42Wor/mbkauthepy"

# --- Explicit Package Discovery ---
# ADDED/MODIFIED SECTION: Tells setuptools where to find the package code
[tool.setuptools.packages.find]
where = ["."]  # Look in the current directory (where pyproject.toml is)
include = ["mbkauthepy"] # Include the mbkauthepy packagey
